Split the never-err re-wrap half of mod_p2 into its own spliced file
